merge-master

+++ polars/polars-core/src/chunked_array/ops/take_random.rs
@@ -10,16 +10,16 @@ use unsafe_unwrap::UnsafeUnwrap;
 macro_rules! take_random_get {
     ($self:ident, $index:ident) => {{
         let (chunk_idx, arr_idx) =
-            crate::utils::index_to_chunked_index($self.chunk_lens.iter().copied(), $index);
-        let arr = $self.chunks.get(chunk_idx);
+            crate::utils::index_to_chunked_index($self.chunk_lens.iter().copied(), $index as u32);
+        let arr = $self.chunks.get(chunk_idx as usize);
         match arr {
             Some(arr) => {
-                if arr.is_null(arr_idx) {
+                if arr.is_null(arr_idx as usize) {
                     None
                 } else {
                     // SAFETY:
                     // bounds checked above
-                    unsafe { Some(arr.value_unchecked(arr_idx)) }
+                    unsafe { Some(arr.value_unchecked(arr_idx as usize)) }
                 }
             }
             None => None,
@@ -30,11 +30,11 @@ macro_rules! take_random_get {
 macro_rules! take_random_get_unchecked {
     ($self:ident, $index:ident) => {{
         let (chunk_idx, arr_idx) =
-            crate::utils::index_to_chunked_index($self.chunk_lens.iter().copied(), $index);
+            crate::utils::index_to_chunked_index($self.chunk_lens.iter().copied(), $index as u32);
         $self
             .chunks
-            .get_unchecked(chunk_idx)
-            .value_unchecked(arr_idx)
+            .get_unchecked(chunk_idx as usize)
+            .value_unchecked(arr_idx as usize)
     }};
 }
 
@@ -146,7 +146,7 @@ where
         } else {
             let t = NumTakeRandomChunked {
                 chunks: chunks.collect(),
-                chunk_lens: self.chunks.iter().map(|a| a.len()).collect(),
+                chunk_lens: self.chunks.iter().map(|a| a.len() as u32).collect(),
             };
             TakeRandBranch3::Multi(t)
         }
@@ -155,7 +155,7 @@ where
 
 pub struct Utf8TakeRandom<'a> {
     chunks: Chunks<'a, LargeStringArray>,
-    chunk_lens: Vec<usize>,
+    chunk_lens: Vec<u32>,
 }
 
 impl<'a> TakeRandom for Utf8TakeRandom<'a> {
@@ -205,7 +205,7 @@ impl<'a> IntoTakeRandom<'a> for &'a Utf8Chunked {
                 let chunks = self.downcast_chunks();
                 let t = Utf8TakeRandom {
                     chunks,
-                    chunk_lens: self.chunks.iter().map(|a| a.len()).collect(),
+                    chunk_lens: self.chunks.iter().map(|a| a.len() as u32).collect(),
                 };
                 TakeRandBranch2::Multi(t)
             }
@@ -228,7 +228,7 @@ impl<'a> IntoTakeRandom<'a> for &'a BooleanChunked {
                 let chunks = self.downcast_chunks();
                 let t = BoolTakeRandom {
                     chunks,
-                    chunk_lens: self.chunks.iter().map(|a| a.len()).collect(),
+                    chunk_lens: self.chunks.iter().map(|a| a.len() as u32).collect(),
                 };
                 TakeRandBranch2::Multi(t)
             }
@@ -252,7 +252,7 @@ impl<'a> IntoTakeRandom<'a> for &'a ListChunked {
             let t = ListTakeRandom {
                 ca: self,
                 chunks: chunks.collect(),
-                chunk_lens: self.chunks.iter().map(|a| a.len()).collect(),
+                chunk_lens: self.chunks.iter().map(|a| a.len() as u32).collect(),
             };
             TakeRandBranch2::Multi(t)
         }
@@ -264,7 +264,7 @@ where
     T: PolarsNumericType,
 {
     chunks: Vec<&'a PrimitiveArray<T>>,
-    chunk_lens: Vec<usize>,
+    chunk_lens: Vec<u32>,
 }
 
 impl<'a, T> TakeRandom for NumTakeRandomChunked<'a, T>
@@ -331,7 +331,7 @@ where
 
 pub struct BoolTakeRandom<'a> {
     chunks: Chunks<'a, BooleanArray>,
-    chunk_lens: Vec<usize>,
+    chunk_lens: Vec<u32>,
 }
 
 impl<'a> TakeRandom for BoolTakeRandom<'a> {
@@ -369,7 +369,7 @@ impl<'a> TakeRandom for BoolTakeRandomSingleChunk<'a> {
 pub struct ListTakeRandom<'a> {
     ca: &'a ListChunked,
     chunks: Vec<&'a LargeListArray>,
-    chunk_lens: Vec<usize>,
+    chunk_lens: Vec<u32>,
 }
 
 impl<'a> TakeRandom for ListTakeRandom<'a> {
@@ -419,7 +419,7 @@ impl<'a> TakeRandom for ListTakeRandomSingleChunk<'a> {
 #[cfg(feature = "object")]
 pub struct ObjectTakeRandom<'a, T: PolarsObject> {
     chunks: Vec<&'a ObjectArray<T>>,
-    chunk_lens: Vec<usize>,
+    chunk_lens: Vec<u32>,
 }
 
 #[cfg(feature = "object")]
@@ -472,7 +472,7 @@ impl<'a, T: PolarsObject> IntoTakeRandom<'a> for &'a ObjectChunked<T> {
         } else {
             let t = ObjectTakeRandom {
                 chunks: chunks.collect(),
-                chunk_lens: self.chunks.iter().map(|a| a.len()).collect(),
+                chunk_lens: self.chunks.iter().map(|a| a.len() as u32).collect(),
             };
             TakeRandBranch2::Multi(t)
         }
